Adds 'addPrivateKey' method to ring signers
