+++ main/signers/hot/SeedSigner/index.js
@@ -10,12 +10,6 @@ class SeedSigner extends HotSigner {
     super(signer)
     this.encryptedSeed = signer.encryptedSeed
     this.worker = fork(WORKER_PATH)
-    setTimeout(() => {
-      this.unlock('frame')
-      setTimeout(() => {
-        this._debug()
-      }, 1000)
-    }, 1000)
   }
 
   save () { super.save({ encryptedSeed: this.encryptedSeed }) }
